Created module i_need_change
This will allow you to call the function whenever you need to, you need to provide a dict of what values you have in the drawer and the count of each value and pass also the change you need. It will what is left in the draw, what you need to give back for each value and if you still owe them money

There is a bug still in play when it comes to decimal points, sometime it will short you a couple of penny's most of the time just a single penny.